# Improve Performance of `read` and `readfiletostring` Functions