This PR significantly improves the performance of the `read` and `readfiletostring` functions in the filesystem library by optimizing how data from `uv_fs_read` is stored.

The functions now use the same approach as `readasync`, storing the read result in a `std::vector<char>` by pushing each character of the `uv_fs_read` result. This replaces the previous method, which appended data to a Luau buffer. The original approach caused frequent buffer resizes, leading to poor performance, especially with large files.

Speed Benchmark:
| File Size | Original | Improved | Speedup |
| --------- | -------- | -------- | ------- |
| 1KB       | 0.0004s  | 0.0001s  | ~4x     |
| 1MB       | 0.3768s  | 0.0037s  | ~102x   |
| 100MB     | 39.4168s | 0.3982s  | ~99x    |

Benchmark Code:
```luau
local fs = require'@lute/fs';

local function benchmark(name, filename)
    local fd = fs.open(filename, 'r');

    local start = os.clock();
    local data = fs.read(fd);
    local took = os.clock() - start;

    fs.close(fd);

    assert(#data ~= 0, 'Error reading: ' .. filename);
    print(string.format('%s %.4f seconds', name, took));
end;

benchmark('1KB   :', '1kb.txt');
benchmark('1MB   :', '1mb.txt');
benchmark('100MB :', '100mb.txt');
```