@SnowyShiro 2025-02-13
Oh yea, are y'all open to people adding stuff to Lute? I wanted to add/refactor some stuff in the filesystem lib, like adding `fs.stat` and `fs.readDir`