@aatxe 2025-02-19
Yeah, totally open to additions to lute. They'll obviously be reviewed for reasonableness, but directory support is pretty much a no-brainer. :)