@aatxe 2025-02-19
Thanks for the improvements, I actually meant to do this kind of thing myself, but you beat me to it! 😸